% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{BoundaryGeometry}
\alias{BoundaryGeometry}
\title{Compute the geometry of the boundaries between areal units}
\usage{
BoundaryGeometry(boundary_list, sf_geometry)
}
\arguments{
\item{boundary_list}{A list of length \mjseqn{I}, where \mjseqn{I} is the number of areal units.
Element \mjseqn{i} of the list contains a vector of indices representing the areal units
that share a boundary with areal unit \mjseqn{i}.}

\item{sf_geometry}{An \code{sf} object representing the geometry of the areal units.}
}
\value{
An \code{sf} object containing the geometry of the boundaries between the areal units.
}
\description{
This utility takes as input a list of boundaries between areal units and the
corresponding \code{sf} geometry object and computes the geometry of the boundaries
between the areal units.
}
