% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ComputeDensities}
\alias{ComputeDensities}
\title{Compute the estimated posterior densities}
\usage{
ComputeDensities(deserialized_chains, N,
                 ranges, names = NULL)
}
\arguments{
\item{deserialized_chains}{A list of \code{RProtoBuf::Message} which stores the deserialized output of
the sampler (either with fuxed or variable number of components).}

\item{N}{An integer representing the number of points of the grid on which the estimated density will be computed.}

\item{ranges}{A matrix of dimesions \mjseqn{2 \times I}, where \mjseqn{I} is the number of areal locations. This
matrix represents the extrema of the grid on which the density estimation will be performed.}

\item{names}{A vector of strings of length \mjseqn{I}, optional input that sets the names of the output elements.
Default value is set to \code{NULL}, thus no name will be attributed.}
}
\value{
A list of \mjseqn{I} elements, where the \mjseqn{i}.th element represent the estimated density over a fixed
grid of \code{N} points between \code{ranges[1,i]} and \code{ranges[2,i]}.
}
\description{
\loadmathjax This utility takes as input the deserialized output of the samplers
(via \code{\link{DeserializeSPMIXProto}}) and compute the posterior estimates of the densities for each
areal location \mjseqn{i=1,\dots,I}.
}
