% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ComputePosteriorLPDF}
\alias{ComputePosteriorLPDF}
\title{Compute the chain of the estimated posterior log-likelihood for each data point}
\usage{
ComputePosteriorLPDF(data, deserialized_chains, verbose = FALSE)
}
\arguments{
\item{data}{The data that needs to be fitted by the model. Data are passed as a list of vectors, whose
\mjseqn{i}-th element represents the vector of data assigned to the \mjseqn{i}-th location.}

\item{deserialized_chains}{A list of \code{RProtoBuf::Message} which stores the deserialized output of
the sampler (either with a fixed or a variable number of components).}

\item{verbose}{A bool. If \code{TRUE}, prints the progress of the computation.}
}
\value{
A \mjseqn{T \times N} matrix, \mjseqn{T} being the number of iterations
of the MCMC chain and \mjseqn{N} the total number of data points.
Element \mjseqn{t,n} of the matrix is the log-likelihood of the \mjseqn{n}-th
data point at \mjseqn{t}-th iteration.
}
\description{
\loadmathjax This utility takes as input the deserialized output of the samplers
(via \code{\link{DeserializeSPMIXProto}}) and compute the chain of posterior log-likelihood
for each data point
}
