% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputePosteriorLPDFs}
\alias{ComputePosteriorLPDFs}
\title{Compute the log posterior densities for each group and for each data point}
\usage{
ComputePosteriorLPDFs(serialized_states, data, display_progress = TRUE)
}
\arguments{
\item{serialized_states}{A list of \code{raw} vectore which stores the serialized output of
the sampler (either with a fixed or a variable number of components).}

\item{data}{The data that needs to be fitted by the model. Data are passed as a list of vectors, whose
\mjseqn{i}-th element represents the vector of data assigned to the \mjseqn{i}-th location.}

\item{display_progress}{A bool. If \code{TRUE}, prints the progress of the computation.}
}
\value{
A \mjseqn{T \times N} matrix, \mjseqn{T} being the number of iterations
of the MCMC chain and \mjseqn{N} the total number of data points.
Element \mjseqn{t,n} of the matrix is the log-likelihood of the \mjseqn{n}-th
data point at \mjseqn{t}-th iteration.
}
\description{
\loadmathjax This utility takes as input the deserialized output of the samplers
(via \code{\link{DeserializeSPMIXProto}}) and compute the chain of posterior log-likelihood
for each data point
}
