% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputePredictiveLPDFs}
\alias{ComputePredictiveLPDFs}
\title{Compute the log predictive densities for each group across a grid of values.}
\usage{
ComputePredictiveLPDFs(serialized_states, grid, display_progress = TRUE)
}
\arguments{
\item{serialized_states}{A list of \code{raw} vectors, where each element is a serialized \code{UnivariateState} protobuf object.
Each serialized state represents a sample from the MCMC chain.}

\item{grid}{An numeric vector representing the grid of values at which to compute the predictive log pdf.}

\item{display_progress}{(Optional) a bool, if \code{TRUE}, it display a progress bar during the computation.}
}
\value{
A list of matrices. Each element of the list is associated to a group and contains a matrix of size
\code{number_of_iterations} times \code{len(grid)}, each element representing the log predictive density for that group
at the corresponding iteration and grid point.
}
\description{
Given the serialized MCMC chain from \code{Sampler.DensityEstimation} or \code{Sampler.BoundaryDetection}, this function
computes the predictive log density in each area over the points specified by \code{grid}.
}
