% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DeserializeSPMIXProto}
\alias{DeserializeSPMIXProto}
\title{Deserialize messages of the \code{SPMIX} package}
\usage{
DeserializeSPMIXProto(message_type, raw_vector)
}
\arguments{
\item{message_type}{A string containing the name of the Message in which the serialized message will be converted}

\item{raw_vector}{A vector of type \code{raw}. The message to be unserialized}
}
\value{
An object of class \code{RProtoBuf::Message}, which stores the unserialized message and can be manipulated
using the \link{RProtoBuf} package.
}
\description{
This funciton is a wrapper to deserialize raw vectors using the \code{.proto} files available in this package.
We rely on Google Protocol Buffers for the serialization procedure and on \code{\link{RProtoBuf}} package to
provide and easy-to-use interface for \code{R} users
}
