% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_ascent_test}
\alias{grad_ascent_test}
\title{Test to evaluate times and correctness of gradient ascent method for optimization on a test function}
\usage{
grad_ascent_test(options)
}
\description{
Test to evaluate times and correctness of gradient ascent method for optimization on a test function
}
