% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hessian_test}
\alias{hessian_test}
\title{Test to use stan math library for automatic differentiation}
\usage{
hessian_test()
}
\description{
Test to use stan math library for automatic differentiation
}
