% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IncreaseMove_test}
\alias{IncreaseMove_test}
\title{Test to evaluate the acceptance rate in case of extension move}
\usage{
IncreaseMove_test(data, W, params, state, options)
}
\description{
Test to evaluate the acceptance rate in case of extension move
}
