% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_alr}
\alias{inv_alr}
\title{Inverse additive log ratio}
\usage{
inv_alr(x, padded_zero = FALSE)
}
\arguments{
\item{x}{Vector of double in \ifelse{html}{\out{R<sup>H-1</sup>}}{\eqn{R^{H-1}}}.}

\item{padded_zero}{Boolean, default to FALSE.}
}
\value{
Vector of double in \ifelse{html}{\out{R<sup>H</sup>}}{\eqn{R^{H}}} i.e.
\ifelse{html}{\out{alr <sup>-1</sup>(x)}}{\eqn{alr^{-1}(x)}}.
}
\description{
This utility computes the inverse additive log ratio transorm of a given vector. This transofrmation is
defined as: (METTI DEF)
}
