% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invAlr}
\alias{invAlr}
\title{Inverse Additive Log Ratio}
\usage{
invAlr(x)
}
\arguments{
\item{x}{Vector of double in \ifelse{html}{\out{R<sup>H-1</sup>}}{\eqn{R^{H-1}}}.}
}
\value{
Vector of double in \ifelse{html}{\out{R<sup>H</sup>}}{\eqn{R^{H}}} i.e.
\ifelse{html}{\out{alr <sup>-1</sup>(x)}}{\eqn{alr^{-1}(x)}}.
}
\description{
This utility computes the inverse additive log ratio transorm of a given vector. This transformation is
defined as: (METTI DEF)
}
