% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{messageFromR}
\alias{messageFromR}
\title{Simple test to see if messages in R can be passed to a C++ function}
\usage{
messageFromR(params)
}
\description{
Simple test to see if messages in R can be passed to a C++ function
}
