% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newton_opt_test}
\alias{newton_opt_test}
\title{Test to evaluate times and correctness of newton method for optimization}
\usage{
newton_opt_test(state, data, W, params, options)
}
\description{
Test to evaluate times and correctness of newton method for optimization
}
