% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protobuf_messages.R
\name{OptimOptions [Protobuf::Message]}
\alias{OptimOptions [Protobuf::Message]}
\title{Options for optimization algorithms}
\format{
A Protobuf::Message with the following fields:
\describe{
\item{max_iter}{An integer storing the maximum number of iterations that the optimization algorithm
have to perform.}
\item{tol}{A double storing the minimum tolerance used to evaluate if the optimization algorithm
has reached convergence.}
}
}
\description{
This protobuf message encodes the options required to the optimization algorithms implemented
in the package and used in the reversible jump sampler. Using \code{\link{RProtoBuf}} package one can easily
retrieve the descriptor and build it from scratch or either parse it from a \code{.asciipb} file.
}
\examples{
# Retrieve message descriptor
RProtoBuf::readProtoFiles(system.file("proto/optimization_option.proto", package = "SPMIX"))
names(OptimOptions)

# Build a new message from scratch
params <- RProtoBuf::new(OptimOptions)

# Read from file (NOT RUN)
params <- RProtoBuf::readASCII(OptimOptions, file("path/to/file.asciipb"))

}
