% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readDataFromCSV}
\alias{readDataFromCSV}
\title{Just a check for output format in R when reading data}
\usage{
readDataFromCSV(filename)
}
\description{
Just a check for output format in R when reading data
}
