% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadMatrixFromCSV}
\alias{ReadMatrixFromCSV}
\title{Import Proximity Matrix from File}
\usage{
ReadMatrixFromCSV(filename)
}
\arguments{
\item{filename}{A string identifying the path to a \code{.csv} file from which the matrix will be read.}
}
\value{
The proximity matrix as a usual \code{R matrix} object.
}
\description{
\loadmathjax This function simply reads the proximity matrix \mjseqn{G} of the Spatial Mixture Model from a \code{.csv} file.
This file should not have columns or row headers and it must be written in the visual form of a matrix
only composed by either \mjseqn{0} or \mjseqn{1}. As assumption, the diagonal of this matrix should be \mjseqn{0}.
}
