% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReduceMove_test}
\alias{ReduceMove_test}
\title{Test to evaluate the acceptance rate in case of reduction move}
\usage{
ReduceMove_test(data, W, params, state, options)
}
\description{
Test to evaluate the acceptance rate in case of reduction move
}
