% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sf_ggmap}
\alias{sf_ggmap}
\title{Transform ggmap object to EPSG:3857}
\usage{
sf_ggmap(map)
}
\arguments{
\item{map}{Output of \code{ggmap::get_map} function}
}
\value{
The same map, but with a bounding box suitable to be plot in \code{ggplot}, i.e., in EPSG:3857
coordinate reference system.
}
\description{
This utility takes as input a \code{ggmap} object and transforms its bounding box
to EPSG:3857 coordinate reference system.
}
