% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mix.1D.R
\name{sp.mix.1D}
\alias{sp.mix.1D}
\title{Estimates a Semiparametric Mixture Density for 1-d data}
\usage{
sp.mix.1D(z, tol = 5e-06, max.iter = 30, doplot = TRUE, thre.localFDR = 0.2)
}
\arguments{
\item{z}{1-dimensional vector z, probit-transformed p-values.}

\item{tol}{Stopping criteria for the EM algorithm.
Optimization stops if maximum absolute difference of is smaller than tol (default: 5.0e-6)}

\item{max.iter}{Maximum number of iterations in the EM algorithm. (default: 30)}

\item{doplot}{If TRUE, draw histogram and fitted lines for estimated density. (default: TRUE)}

\item{thre.localFDR}{false positive rate(FDR) and sensitivity with the threshold set to be fdr<=thre.localFDR. (default: 0.2)}
}
\value{
Parametrization of f(x) in terms of hyperplanes and function
  evaluations y = log(f(x)) \item{aOpt, bOpt}{Analytically normalized
  parameters of f(x).} \item{logLike}{Log-likelihood of f(x)} \item{y}{Vector
  with values y_i = log(f(X_)) of the normalized density (\eqn{logLike =
  \sum(y_i)}).} \item{aOptSparse, bOptSparse}{Sparse parametrization
  normalized on the integration grid.}
}
\description{
\code{sp.mix.1D} returns a semiparametric mixture density estimates for given 1-d z, which are the probit-transformed p-values.
}
