% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mix.multi.R
\name{sp.mix.multi}
\alias{sp.mix.multi}
\title{LocalFDR estimation for multi-dimensional lists of z-values}
\usage{
sp.mix.multi(z, tol = 5e-06, max.iter = 30, mono = TRUE)
}
\arguments{
\item{z}{Matrix which column indicates z-values, probit-transformed p-values.}

\item{tol}{Stopping criteria for the EM algorithm. If maximum absolute difference of current and previous gamma value is smaller than tol, i.e. \eqn{max_i |\gamma_i^{(k+1)}-\gamma_i^{(k)} <tol}, for k-th step, then optimization stops. (default: 5e-6)}

\item{max.iter}{Maximum number of iterations in the EM algorithm. (default: 30)}

\item{mono}{If TRUE, LocalFDR is in ascending order of z-values. (default: TRUE)}
}
\value{
Parametrization of f(x) in terms of hyperplanes and function
  evaluations y = log(f(x)) \item{aOpt, bOpt}{Analytically normalized
  parameters of f(x).} \item{logLike}{Log-likelihood of f(x)} \item{y}{Vector
  with values y_i = log(f(X_)) of the normalized density (\eqn{logLike =
  \sum(y_i)}).} \item{aOptSparse, bOptSparse}{Sparse parametrization
  normalized on the integration grid.}
}
\description{
\code{sp.mix.multi} returns LocalFDR estimates and semiparametric mixture density estimates for given multi-dimensional lists of z-values, which are the probit-transformed p-values.
For the hypothesis testing \code{sp.mix.multi} uses a two-component semiparametric mixture model to estimate the LocalFDR from the p-values. The two pillars of the proposed approach are Efron’s empirical null principle and log-concave density estimation for the alternative distribution.
Estimated prior, null and alternative probability for given points and estimated LocalFDR are returned.
}
