% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPMIX_SamplerFromFiles}
\alias{SPMIX_SamplerFromFiles}
\title{First try on the spatial sampler, taking input from files}
\usage{
SPMIX_SamplerFromFiles(burnin, niter, thin, data_file, w_file, params_file)
}
\description{
First try on the spatial sampler, taking input from files
}
