% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spmixLogLikelihood}
\alias{spmixLogLikelihood}
\title{Loglikelihood of a Spatial Mixture model state}
\usage{
spmixLogLikelihood(state, params)
}
\description{
Loglikelihood of a Spatial Mixture model state
}
