% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_spmixloglikelihood}
\alias{test_spmixloglikelihood}
\title{Test for loglikelohood with m_tilde contribution}
\usage{
test_spmixloglikelihood(data, W, params_str, state_str)
}
\description{
Test for loglikelohood with m_tilde contribution
}
