% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{unserialize_proto}
\alias{unserialize_proto}
\title{Unserialize raw vectors}
\usage{
unserialize_proto(message_type, raw_vector)
}
\arguments{
\item{message_type}{A string containing the name of the Message in which the serialized message will be converted}

\item{raw_vector}{A vector of type \code{raw}. The message to be unserialized}
}
\value{
An object of class \code{RProtoBuf::Message}, which stores the unserialized message and can be manipulated
using \link{RProtoBuf}.
}
\description{
This funciton is a wrapper to unserialize raw vectors that comes from the serialization of
messages using Google Protocol Buffer API using the proto files available in the \code{SPMIX} package.
}
