% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{unserializeSPMIXProto}
\alias{unserializeSPMIXProto}
\title{Unserialize Messages of the \code{SPMIX} Package}
\usage{
unserializeSPMIXProto(message_type, raw_vector)
}
\arguments{
\item{message_type}{A string containing the name of the Message in which the serialized message will be converted}

\item{raw_vector}{A vector of type \code{raw}. The message to be unserialized}
}
\value{
An object of class \code{RProtoBuf::Message}, which stores the unserialized message and can be manipulated
using \link{RProtoBuf} package.
}
\description{
This funciton is a wrapper to unserialize raw vectors using the \code{.proto} files available in this package.
We rely on Google Protocol Buffer for the serialization procedure and on \code{\link{RProtoBuf}} package to
provide and easy-to-use interface for \code{R} users
}
