% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alr}
\alias{alr}
\title{Additive log ratio}
\usage{
alr(x, pad_zero = FALSE)
}
\arguments{
\item{x}{Vector of double in \ifelse{html}{\out{R<sup>H</sup>}}{\eqn{R^{H}}}.}

\item{pad_zero}{Boolean, default to FALSE.}
}
\value{
Vector of double in \ifelse{html}{\out{R<sup>H-1</sup>}}{\eqn{R^{H-1}}} i.e. alr(x).
}
\description{
This utility computes the additive log ratio transorm of a given vector. This transformation is
defined as: (METTI DEF)
}
