% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alr}
\alias{alr}
\title{Additive Log Ratio}
\usage{
alr(x)
}
\arguments{
\item{x}{Vector of double in the simplex \mjseqn{S^H}.}
}
\value{
Vector of double in \mjseqn{\mathbb{R}^{H-1}} i.e. \mjseqn{\operatorname{alr}(x)}.
}
\description{
\loadmathjax This utility computes the additive log ratio transorm of a given vector. Given a generic vector of the simplex
\mjseqn{x \in S^{H}}, the transformation is defined as:
\mjsdeqn{ \operatorname{alr}(x)_j = \log \left( \frac{x_j}{x_H} \right) \quad \forall j=1,\dots,H-1 }
}
