% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ComputeDensities}
\alias{ComputeDensities}
\title{Compute the chain of the estimated posterior densities over a grid of points}
\usage{
ComputeDensities(deserialized_chains, x_grid, verbose = FALSE)
}
\arguments{
\item{deserialized_chains}{A list of \code{RProtoBuf::Message} which stores the deserialized output of
the sampler (either with a fixed or a variable number of components).}

\item{x_grid}{A numeric vector representing the grid of points on which the estimated densities will be evaluated.}

\item{verbose}{A bool. If \code{TRUE}, prints the progress of the computation.}
}
\value{
A list of \mjseqn{I} elements, where the \mjseqn{i}-th element is a matrix that represent, by row,
the estimated density, the lower and the upper bounds the corresponding credibility interval
(if such quantity is required) evaluated over \code{x_grid}.
}
\description{
\loadmathjax This utility takes as input the deserialized output of the samplers
(via \code{\link{DeserializeSPMIXProto}}) and compute the chain of posterior estimates of the densities for each
areal location \mjseqn{i=1,\dots,I}.
}
