% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InvAlr}
\alias{InvAlr}
\title{Inverse Additive Log Ratio}
\usage{
InvAlr(x)
}
\arguments{
\item{x}{Vector of double in \mjseqn{ \mathbb{R}^{H-1} }.}
}
\value{
Vector of double in the simplex \mjseqn{S^H} i.e. \mjseqn{\operatorname{alr}^{-1}(x)}.
}
\description{
\loadmathjax This utility computes the inverse additive log ratio transform of a given vector. Given a generic vector
\mjseqn{ x \in \mathbb{R}^{H-1} }, the transformation is defined as:
\mjsdeqn{ \begin{eqnarray*}
\operatorname{alr}^{-1}(x)_{j} &=& \textstyle\frac{\operatorname{e}^{x_j}}{\sum _{h} \operatorname{e}^{x_h}} \quad \forall j=1,\dots,H-1 \cr
\operatorname{alr}^{-1}(x)_H &=& 1 - \textstyle\sum _{h} \operatorname{alr}^{-1}(x)_h
\end{eqnarray*} }
}
