% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadDataFromCSV}
\alias{ReadDataFromCSV}
\title{Import Data from File}
\usage{
ReadDataFromCSV(filename)
}
\value{
A list of dimension \mjseqn{I}, in which the \mjseqn{i}-th element is a vector containing all data that
have been assigned to the \mjseqn{i}-th location.
}
\description{
\loadmathjax This utility reads the input data for the sampler from a \code{.csv} file. This file needs to be provided
with column headers. Moreover, it should be organized in two columns for the correct parsing:
\itemize{
\item{\strong{group}, an integer (from \mjseqn{0} to \mjseqn{I-1}) describing the area to which the data belongs to;}
\item{\strong{data}, the actual data, which will be assigned to area indicated by the "group" column.}
}
}
