% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readMatrixFromCSV}
\alias{readMatrixFromCSV}
\title{Just a check for output format in R when reading a matrix}
\usage{
readMatrixFromCSV(filename)
}
\description{
Just a check for output format in R when reading a matrix
}
