% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-linearRegression.R
\name{Sampler.LinearRegression}
\alias{Sampler.LinearRegression}
\title{\code{SPMIX} Sampler for Linear Regression}
\usage{
Sampler.LinearRegression(
  burnin,
  niter,
  thin,
  data,
  W,
  params,
  cov = list(),
  options = NULL,
  display_progress = TRUE,
  seed = NULL
)
}
\arguments{
\item{burnin}{Integer, the number of steps of the burnin phase.}

\item{niter}{Integer, the number of steps to run \emph{after} the burnin.}

\item{thin}{Integer, is the thinning parameter, hence only one every 'thin' iterations will be saved
during the sampling phase.}

\item{data}{The data that needs to be fitted by the model. Data can be passed either as string, thus indicating
the path to a \code{.csv} file representing the data, or as a list of vectors. In this case, the
\mjseqn{i}-th element of the data list represents the vector of data assigned to the \mjseqn{i}-th location.}

\item{W}{The proximity matrix between different locations. It can be passed as string, i.e. a string containing
the path to a \code{.csv} file storing the proximity matrix, or as a simple R matrix.}

\item{params}{The sampler parameters that needs to be provided as input. params can be passed as a string
containing the path to an \code{.asciipb} file or as an \code{S4::Message} class of type SamplerParams, i.e. a
Google Protocol Buffer Message available in the package and interfaced to R through \code{\link{RProtoBuf}}.}

\item{cov}{A list of vectors that represents covariates. As default value, it is an empty list.
In case this input parameter is not empty, the sampler performs a regression on these covariates.}

\item{options}{The sampler optimization options used in the execution of the reversible jump sampler.
Default value is set to \code{NULL} and in case type is "rjmcmc", a \code{S4::Message} object of type
OptimOptions is istanciated with default values. In order to override the default values, options can be
passed either as a string, representing the path to an \code{.asciipb} file, or as a \code{S4::Message}
object of type OptimOptions generated via \code{\link{RProtoBuf}}.}

\item{display_progress}{Boolean, it allows you display on the console the progress bar during burn-in
and sampling phase. As default value, it is set to TRUE.}

\item{seed}{Integer, the random seed to be used in the sampler. If set to NULL (default value),
the seed is generated using the current system time.}
}
\value{
A list of raw vectors where the \mjseqn{i}-th element is the \mjseqn{i}-th saved draw.
In order to reduce the space occupied by these draws, data are serialized using Google Protocol Buffers.
Each state can be easily deserialized in R using the \code{\link{DeserializeSPMIXProto}} function of this package.
}
\description{
\loadmathjax Runs the Gibbs samplers for the \code{SPMIX} model for a total of burnin + niter iterations,
discarding the first 'burnin' ones and keeping in memory only one every 'thin' iterations.
}
