% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mix.multi.R
\name{sp.mix.multi}
\alias{sp.mix.multi}
\title{Estimates a Semiparametric Mixture Density for multi-dimensional data}
\usage{
sp.mix.multi(z, tol = 5e-06, max.iter = 30, mono = TRUE)
}
\arguments{
\item{z}{Matrix which each column indicates z, probit-transformed p-values.}

\item{tol}{Stopping criteria for the EM algorithm.
Optimization stops if maximum absolute difference of current and previous gamma value is smaller than tol. (default: 5.0e-6)}

\item{max.iter}{Maximum number of iterations in the EM algorithm. (default: 30)}

\item{mono}{If TRUE, monotone. (default: TRUE)}
}
\value{
Parametrization of f(x) in terms of hyperplanes and function
  evaluations y = log(f(x)) \item{aOpt, bOpt}{Analytically normalized
  parameters of f(x).} \item{logLike}{Log-likelihood of f(x)} \item{y}{Vector
  with values y_i = log(f(X_)) of the normalized density (\eqn{logLike =
  \sum(y_i)}).} \item{aOptSparse, bOptSparse}{Sparse parametrization
  normalized on the integration grid.}
}
\description{
\code{sp.mix.multi} returns a semiparametric mixture density estimates for given multi-dimensional z, which are the probit-transformed p-values.
}
