% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SPMIX_sampler}
\alias{SPMIX_sampler}
\title{Spatial Mixture Model Sampler}
\usage{
SPMIX_sampler(
  burnin,
  niter,
  thin,
  data,
  W,
  params,
  cov = list(),
  display_progress = TRUE
)
}
\arguments{
\item{burnin}{blabla}

\item{niter}{blabla}

\item{thin}{blabla}

\item{data}{blabla}

\item{W}{blabla}

\item{params}{blabla}

\item{cov}{blabla}

\item{display_progress}{blabla}
}
\value{
bleble
}
\description{
This function provides the sampling algorithm for Spatial Mixture Models data. A seguire, se tutto funziona,
una bella documentazione.
}
