% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{SPMIX_sampler}
\alias{SPMIX_sampler}
\title{Spatial Mixture Model Sampler}
\usage{
SPMIX_sampler(burnin, niter, thin, data, W, params,
    cov = list(), type = "no_rjmcmc", options = NULL, display_progress = TRUE)
}
\arguments{
\item{burnin}{Integer, the number of steps of the burnin phase.}

\item{niter}{Integer, the number of steps to run \emph{after} the burnin,}

\item{thin}{Integer, is the thinning parameter, hence only one every 'thin' iterations will be saved
during the sampling phase.}

\item{data}{The data that needs to be fitted by the model. Data can be passed either as string, thus indicating
the path to a \code{.csv} file representing the data, or as a list of vectors. In this case, i-th data represents
the vector of data in location i.}

\item{W}{The proximity matrix between different locations. W can be passed as string, i.e. a string containing
the path to a \code{.csv} file storing the proximity matrix, or as a simple R matrix.}

\item{params}{The sampler parameters that needs to be provided as input. params can be passed as a string
containing the path to an \code{.asciipb} file or as an \code{S4::Message} class of type SamplerParams, i.e. a
Google Protocol Buffer Message available in the package and interfaced to R through \code{\link{RProtoBuf}}.}

\item{cov}{A list of vectors that represents covariates. As default value, is an empty list.
In case this input parameter is not empty, the sampler performs a regression on these covariates.}

\item{type}{A string identifying the type of sampler to run. If type is "rjmcmc", the algorithm will run
the spatial mixture sampler putting a prior on the number of components H. Default value is "no_rjmcmc", which
samples from the spatial mixture model with fixed number of components.}

\item{options}{The sampler optimization options used in the execution of the reversible jump sampler.
Default value is set to \code{NULL} and in case type is "rjmcmc", a \code{S4::Message} object of type
OptimOptions is istanciated with default values. In order to overwrite default values, options can be either
a string, representing the path to an \code{.asciipb} file or a \code{S4::Message} object of type OptimOptions
generated via \code{\link{RProtoBuf}}.}

\item{display_progress}{Boolean, it allows you display on the console the progress bar during burn-in
and sampling phase. As default value, is set to TRUE.}
}
\value{
A list of raw vectors where the i-th element is the i-th saved draw In order to reduce the space
occupied by these draws, data are serialized through Google Protocol Buffers serialization procedure.
Each state can be easily deserialized in R using the \code{\link{unserialize_proto}} function in this package.
}
\description{
Runs the Gibbs sampler for the SPMIX model for a total of burnin + niter iterations,
discarding the first 'burnin' ones and keeping in memory only one every 'thin' iterations.
}
